backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pyttsx3==2.90
pandas==1.3.3
nltk==3.6.3
easyocr==1.4.1
scikit-learn==0.24.2
sentence-transformers==2.2.2
Pillow==8.3.2
numpy==1.21.2